backend/requirements.txt
fastapi==0.121.2
uvicorn[standard]==0.38.0
python-dotenv==1.2.1
pydantic==2.12.4
pydantic-settings==2.12.0
replicate==1.0.7
openai==2.8.0
ffmpeg-python==0.2.0
httpx==0.28.1
python-multipart==0.0.20
modal>=0.63.0
Pillow>=10.0.0
requests>=2.31.0
pytest>=7.4.4
pytest-asyncio>=0.21.0